train_model/requirements.txt
pandas
scikit-learn==1.3.2
pickledb
psycopg2-binary==2.9.6
SQLAlchemy==2.0.20
